add better styling for break
